[SBL-199] 설문 Management 페이지 관련 API들을 SurveyManagementController로 이동
