+++ src/main/kotlin/com/sbl/sulmun2yong/survey/controller/SurveyManagementController.kt
@@ -1,10 +1,11 @@
 package com.sbl.sulmun2yong.survey.controller
 
 import com.sbl.sulmun2yong.global.annotation.LoginUser
-import com.sbl.sulmun2yong.survey.controller.doc.SurveyResultApiDoc
+import com.sbl.sulmun2yong.survey.controller.doc.SurveyManagementApiDoc
 import com.sbl.sulmun2yong.survey.dto.request.SurveyResultRequest
-import com.sbl.sulmun2yong.survey.service.SurveyResultService
+import com.sbl.sulmun2yong.survey.service.SurveyManagementService
 import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
@@ -14,15 +15,21 @@ import org.springframework.web.bind.annotation.RestController
 import java.util.UUID
 
 @RestController
-@RequestMapping("/api/v1/surveys/result")
-class SurveyResultController(
-    private val surveyResultService: SurveyResultService,
-) : SurveyResultApiDoc {
-    @PostMapping("/{survey-id}")
+@RequestMapping("/api/v1/surveys/management")
+class SurveyManagementController(
+    private val surveyManagementService: SurveyManagementService,
+) : SurveyManagementApiDoc {
+    @PostMapping("/result/{survey-id}")
     override fun getSurveyResult(
         @PathVariable("survey-id") surveyId: UUID,
         @LoginUser id: UUID,
         @RequestBody surveyResultRequest: SurveyResultRequest,
         @RequestParam participantId: UUID?,
-    ) = ResponseEntity.ok(surveyResultService.getSurveyResult(surveyId, id, surveyResultRequest, participantId))
+    ) = ResponseEntity.ok(surveyManagementService.getSurveyResult(surveyId, id, surveyResultRequest, participantId))
+
+    @GetMapping("/participants/{survey-id}")
+    override fun getSurveyParticipants(
+        @PathVariable("survey-id") surveyId: UUID,
+        @LoginUser userId: UUID,
+    ) = ResponseEntity.ok(surveyManagementService.getSurveyParticipants(surveyId, userId))
 }
