+++ src/main/kotlin/com/sbl/sulmun2yong/survey/service/SurveyManagementService.kt
@@ -1,18 +1,21 @@
 package com.sbl.sulmun2yong.survey.service
 
+import com.sbl.sulmun2yong.drawing.adapter.DrawingHistoryAdapter
 import com.sbl.sulmun2yong.survey.adapter.ParticipantAdapter
 import com.sbl.sulmun2yong.survey.adapter.ResponseAdapter
 import com.sbl.sulmun2yong.survey.adapter.SurveyAdapter
 import com.sbl.sulmun2yong.survey.dto.request.SurveyResultRequest
+import com.sbl.sulmun2yong.survey.dto.response.ParticipantsInfoListResponse
 import com.sbl.sulmun2yong.survey.dto.response.SurveyResultResponse
 import org.springframework.stereotype.Service
 import java.util.UUID
 
 @Service
-class SurveyResultService(
+class SurveyManagementService(
     private val responseAdapter: ResponseAdapter,
     private val surveyAdapter: SurveyAdapter,
     private val participantAdapter: ParticipantAdapter,
+    private val drawingHistoryAdapter: DrawingHistoryAdapter,
 ) {
     fun getSurveyResult(
         surveyId: UUID,
@@ -39,5 +42,15 @@ class SurveyResultService(
             }
 
         return SurveyResultResponse.of(filteredSurveyResult, survey, participantCount)
+    }
+
+    fun getSurveyParticipants(
+        surveyId: UUID,
+        makerId: UUID,
+    ): ParticipantsInfoListResponse {
+        val survey = surveyAdapter.getByIdAndMakerId(surveyId, makerId)
+        val participants = participantAdapter.findBySurveyId(surveyId)
+        val drawingHistories = if (survey.isImmediateDraw()) drawingHistoryAdapter.getBySurveyId(surveyId, false) else null
+        return ParticipantsInfoListResponse.of(participants, drawingHistories, survey.rewardSetting.targetParticipantCount)
     }
 }
